# packages/openapi/skill-registry.openapi.yaml
openapi: 3.1.0
info:
  title: ManagedSkillHub API
  description: Self-hosted skill registry for AI agents.
  version: 0.1.0
servers:
- url: http://localhost:3040
  description: Local API server
...
paths:
...
  /admin/judge/skill/{skillId}/version/{version}:
    post:
      operationId: judgeSkillVersion
      tags:
      - Judgements
      summary: Judge a published skill version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Judgement result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgementResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /admin/skills:
    get:
      operationId: adminListSkills
      tags:
      - Admin Skills
      summary: List skills including unpublished latest versions
      security:
      - adminSession: []
      responses:
        '200':
          description: Skill list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      operationId: adminCreateSkill
      tags:
      - Admin Skills
      summary: Create a skill from a proposal or draft
      security:
      - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateSkillRequest'
      responses:
        '201':
          description: Created initial draft version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillMutationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}:
    get:
      operationId: adminGetSkill
      tags:
      - Admin Skills
      summary: Get any skill including unpublished versions
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Skill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      operationId: adminUpdateSkill
      tags:
      - Admin Skills
      summary: Update skill draft metadata by creating a new draft version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateSkillRequest'
      responses:
        '200':
          description: Created draft version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillMutationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/skills/{skillId}/files:
    get:
      operationId: adminListSkillFiles
      tags:
      - Admin Skills
      summary: List files for any skill version including unpublished ones
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: string
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillFileListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      operationId: adminUploadSkillFile
      tags:
      - Admin Skills
      summary: Create a new draft version by uploading or replacing a skill file
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                path:
                  type: string
                role:
                  type: string
                  enum:
                  - entrypoint
                  - example
                  - knowledge
                  - test
                  - attachment
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Created draft version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillMutationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}/files/{fileId}:
    get:
      operationId: adminGetSkillFile
      tags:
      - Admin Skills
      summary: Download any skill file including unpublished versions
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      operationId: adminMoveSkillFile
      tags:
      - Admin Skills
      summary: Create a new draft version by moving or renaming a skill file
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminMoveSkillFileRequest'
      responses:
        '200':
          description: Created draft version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillMutationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      operationId: adminDeleteSkillFile
      tags:
      - Admin Skills
      summary: Create a new draft version by removing a skill file
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Created draft version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillMutationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}/files/{fileId}/content:
    put:
      operationId: adminUpdateSkillFileContent
      tags:
      - Admin Skills
      summary: Create a new draft version by replacing text-based file content
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateSkillFileContentRequest'
      responses:
        '200':
          description: Created draft version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillMutationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}/files/{fileId}/extracted-content:
    get:
      operationId: adminGetSkillFileExtractedContent
      tags:
      - Admin Skills
      summary: Get extracted text for any skill file including unpublished versions
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Extracted file content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractedContentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}/files/{fileId}/probe:
    get:
      operationId: adminGetSkillFileProbe
      tags:
      - Admin Skills
      summary: Probe a skill file with ffprobe
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Probe output and parsed summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}/submit-review:
    post:
      operationId: adminSubmitForReview
      tags:
      - Admin Skills
      summary: Submit skill for review
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/skills/{skillId}/approve:
    post:
      operationId: adminApproveSkill
      tags:
      - Admin Skills
      summary: Approve skill version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/skills/{skillId}/publish:
    post:
      operationId: adminPublishSkill
      tags:
      - Admin Skills
      summary: Publish approved skill version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                judgementOverrideReason:
                  type: string
                  minLength: 1
                  description: Audited administrator-only reason for overriding a
                    required judgement policy.
      responses:
        '200':
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/skills/{skillId}/reject:
    post:
      operationId: adminRejectSkill
      tags:
      - Admin Skills
      summary: Reject a draft, in-review, or approved skill version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - reason
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}/deprecate:
    post:
      operationId: adminDeprecateSkill
      tags:
      - Admin Skills
      summary: Deprecate published skill version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Deprecated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/skills/{skillId}/files/{fileId}/re-extract:
    post:
      operationId: adminReextractSkillFile
      tags:
      - Admin Skills
      summary: Re-run extraction for a skill file without modifying the original artifact
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Extracted content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractedContentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/skills/{skillId}/versions/{version}/re-judge:
    post:
      operationId: adminRejudgeSkillVersion
      tags:
      - Admin Skills
      summary: Re-run judgement for a skill version
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Judgement result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JudgementResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/search/reindex:
    post:
      operationId: adminReindexSearch
      tags:
      - Admin Skills
      summary: Rebuild the published skill search index
      security:
      - adminSession: []
      responses:
        '200':
          description: Reindex result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReindexSearchResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /admin/projections/rebuild:
    post:
      operationId: adminRebuildProjections
      tags:
      - Admin Skills
      summary: Rebuild relational projections for configured catalog/search providers
      security:
      - adminSession: []
      parameters:
      - name: clearProjections
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Rebuild summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebuildProjectionsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: skill_hub_session
    publicReadBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for published skill consumption
        when PUBLIC_READ_AUTH_MODE=bearer.
    proposalBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for proposal duplicate check,
        submit, upload, finalization, notices, and status when PROPOSAL_AUTH_MODE=bearer.
    discoveryBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for discovery/contract endpoints
        when DISCOVERY_AUTH_MODE=bearer.
    agentOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.example.invalid/application/o/managedskillhub-agent-device/.well-known/openid-configuration
      description: Runtime Authentik access token. The concrete issuer, client ID,
        scopes, Device Authorization endpoint, token endpoint, and OpenID configuration
        URL must be read from /discover; this static URL is intentionally only a contract
        placeholder.
      x-device-authorization:
        flow: device_code
        runtimeMetadataPath: /discover
        verificationLinkField: verification_uri_complete
        deviceCodeIsSecret: true
    agentSession:
      type: apiKey
      in: header
      name: Authorization
      description: 'Short-lived agent session code sent as "Authorization: AgentSession
        <code>". Created through POST /agent-sessions after proving possession of
        the relevant area bearer tokens.'
...
  schemas:
...
    JudgementResponse:
      type: object
      required:
      - id
      - targetType
      - targetId
      - dimensions
      - overallRisk
      - summary
      - skillPurposeSummary
      - createdAt
      properties:
        id:
          type: string
        targetType:
          type: string
          enum:
          - proposal
          - skill
          - file
        targetId:
          type: string
        dimensions:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/JudgementDimension'
        overallRisk:
          type: string
          enum:
          - low
          - medium
          - high
          - critical
          - no_judge_available
        summary:
          type: string
        skillPurposeSummary:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...